fix: optimize b3CheckPermissions utils
